Handle Substrait's `protobuf` module rename
